Migrated tests to separate file
